Litt mer rydding i metoder og klasser
